backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
asyncpg
langchain
langchain-core
langchain-openai
langgraph
python-dotenv
pydantic
paramiko
ansible
sse-starlette
websockets
aiohttp
requests
llama-api-client
pytest
pytest-asyncio
greenlet